Move help preferences to confpage.py
